requirements.txt
flask
gunicorn
python-dotenv
groq
requests